+++ packages/runtime-ts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @orcalang/orca-runtime-ts
 
+## 0.1.12
+
+### Patch Changes
+
+- chore: version bump
+
+---
+
 ## 0.1.11
 
 ### Patch Changes